requirements.txt
fastapi
uvicorn
firebase-admin
python-dotenv
pydantic
pytest
